save, delete tasks add

+++ ProjetoTasks/src/main/java/com/example/ProjetoTasks/Services/TaskService.java
@@ -45,6 +45,10 @@ public class TaskService {
 
     @Transactional
     public void deleteTask(Long id){
+
+        if(!taskRepository.existsById(id)){
+            throw new RuntimeException("Task not found.");
+        }
         taskRepository.deleteById(id);
     }
 }
